@@ -88,7 +88,7 @@
         <rest-assured.version>3.3.0</rest-assured.version>
         <!-- plugins -->
         <thin.version>1.0.22.RELEASE</thin.version>
-        <spring-boot.version>2.6.4</spring-boot.version>
+        <spring-boot.version>2.7.2</spring-boot.version>
         <aspectjweaver.version>1.9.1</aspectjweaver.version>
     </properties>
 
