@@ -7,19 +7,21 @@ import org.paukov.combinatorics3.PermutationGenerator.TreatDuplicatesAs;
 
 public class StringPermutationsCombinatoricsLib {
 
-    public List<List<Character>> permutationWithoutRepetitions(final String string) {
-        List<Character> chars = ArrayHelper.toCharacterList(string);
-        return Generator.permutation(chars)
-            .simple()
-            .stream()
-            .collect(Collectors.toList());
-    }
+public List<String> permutationWithoutRepetitions(final String string) {
+    List<Character> chars = Helper.toCharacterList(string);
+    return Generator.permutation(chars)
+        .simple()
+        .stream()
+        .map(Helper::toString)
+        .collect(Collectors.toList());
+}
 
-    public List<List<Character>> permutationWithRepetitions(final String string) {
-        List<Character> chars = ArrayHelper.toCharacterList(string);
+    public List<String> permutationWithRepetitions(final String string) {
+        List<Character> chars = Helper.toCharacterList(string);
         return Generator.permutation(chars)
             .simple(TreatDuplicatesAs.IDENTICAL)
             .stream()
+            .map(Helper::toString)
             .collect(Collectors.toList());
     }
 
