@@ -2,22 +2,19 @@ package com.baeldung.algorithms.stringpermutation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Collection;
 import java.util.List;
-import org.apache.commons.collections4.iterators.PermutationIterator;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 class StringPermutationsApacheUnitTest {
 
+    @ParameterizedTest
     @CsvSource({"abc, 6",
                 "hello, 120",
                 "aaaaaa, 720"})
-    @DisplayName("Apache permutation for ")
     void testPermutationsWithRepetitions(String string, int numberOfPermutations) {
         StringPermutationsApache permutationGenerator = new StringPermutationsApache();
-        final Collection<List<Character>> permutations = permutationGenerator.eagerPermutationWithRepetitions(string);
+        final List<String> permutations = permutationGenerator.eagerPermutationWithRepetitions(string);
         final int size = permutations.size();
         assertThat(permutations)
             .as("\"%s\" should have %d permutation, but had %d", string, numberOfPermutations, size)
@@ -30,12 +27,8 @@ class StringPermutationsApacheUnitTest {
         "aaaaaa, 720"})
     void testPermutationsWithoutRepetitions(String string, int numberOfPermutations) {
         StringPermutationsApache permutationGenerator = new StringPermutationsApache();
-        final PermutationIterator<Character> permutations = permutationGenerator.lazyPermutationWithoutRepetitions(string);
-        int size = 0;
-        while (permutations.hasNext()) {
-            permutations.next();
-            ++size;
-        }
+        final List<String> permutations = permutationGenerator.lazyPermutationWithoutRepetitions(string);
+        int size = permutations.size();
         assertThat(size)
             .as("\"%s\" should have %d permutation, but had %d", string, numberOfPermutations, size)
             .isEqualTo(numberOfPermutations);
