@@ -14,7 +14,7 @@ class StringPermutationsCombinatoricsLibUnitTest {
                 "aaaaaa, 720"})
     void testPermutationsWithRepetitions(String string, int numberOfPermutations) {
         StringPermutationsCombinatoricsLib permutationGenerator = new StringPermutationsCombinatoricsLib();
-        final List<List<Character>> permutations = permutationGenerator.permutationWithRepetitions(string);
+        final List<String> permutations = permutationGenerator.permutationWithRepetitions(string);
         final int size = permutations.size();
         assertThat(permutations)
             .as("\"%s\" should have %d permutation, but had %d", string, numberOfPermutations, size)
@@ -27,7 +27,7 @@ class StringPermutationsCombinatoricsLibUnitTest {
                 "aaaaaa, 1"})
     void testPermutationsWithoutRepetitions(String string, int numberOfPermutations) {
         StringPermutationsCombinatoricsLib permutationGenerator = new StringPermutationsCombinatoricsLib();
-        final List<List<Character>> permutations = permutationGenerator.permutationWithoutRepetitions(string);
+        final List<String> permutations = permutationGenerator.permutationWithoutRepetitions(string);
         final int size = permutations.size();
         assertThat(permutations)
             .as("\"%s\" should have %d permutation, but had %d", string, numberOfPermutations, size)
