@@ -2,7 +2,6 @@ package com.baeldung.algorithms.stringpermutation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Collection;
 import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -15,7 +14,7 @@ class StringPermutationsGuavaUnitTest {
         "aaaaaa, 720"})
     void testPermutationsWithRepetitions(String string, int numberOfPermutations) {
         StringPermutationsGuava permutationGenerator = new StringPermutationsGuava();
-        final Collection<List<Character>> permutations = permutationGenerator.permutationWithRepetitions(string);
+        final List<String> permutations = permutationGenerator.permutationWithRepetitions(string);
         final int size = permutations.size();
         assertThat(permutations)
             .as("\"%s\" should have %d permutation, but had %d", string, numberOfPermutations, size)
@@ -28,7 +27,7 @@ class StringPermutationsGuavaUnitTest {
         "aaaaaa, 1"})
     void testPermutationsWithoutRepetitions(String string, int numberOfPermutations) {
         StringPermutationsGuava permutationGenerator = new StringPermutationsGuava();
-        final Collection<List<Character>> permutations = permutationGenerator.permutationWithoutRepetitions(string);
+        final List<String> permutations = permutationGenerator.permutationWithoutRepetitions(string);
         final int size = permutations.size();
         assertThat(permutations)
             .as("\"%s\" should have %d permutation, but had %d", string, numberOfPermutations, size)
