@@ -7,6 +7,6 @@ public class JarAppUnitTest {
 
     @Test
     public void findClassTest(){
-        Assert.assertTrue(JarApp.findObjectMapperClass().endsWith("jackson-databind-2.13.0.jar"));
+        Assert.assertTrue(JarApp.findObjectMapperClass().endsWith("jackson-databind-2.13.3.jar"));
     }
 }
