@@ -3,7 +3,7 @@ package com.baeldung.optaplanner;
 import org.optaplanner.core.api.domain.solution.PlanningEntityCollectionProperty;
 import org.optaplanner.core.api.domain.solution.PlanningScore;
 import org.optaplanner.core.api.domain.solution.PlanningSolution;
-import org.optaplanner.core.api.domain.solution.drools.ProblemFactCollectionProperty;
+import org.optaplanner.core.api.domain.solution.ProblemFactCollectionProperty;
 import org.optaplanner.core.api.domain.valuerange.ValueRangeProvider;
 import org.optaplanner.core.api.score.buildin.hardsoft.HardSoftScore;
 import org.slf4j.Logger;
@@ -16,11 +16,17 @@ import java.util.List;
 public class CourseSchedule {
 
     Logger logger = LoggerFactory.getLogger("CourseSchedule");
-
+    @ValueRangeProvider(id = "availableRooms")
+    @ProblemFactCollectionProperty
     private List<Integer> roomList;
+    @ValueRangeProvider(id = "availablePeriods")
+    @ProblemFactCollectionProperty
     private List<Integer> periodList;
+    @PlanningEntityCollectionProperty
     private List<Lecture> lectureList;
+    @PlanningScore
     private HardSoftScore score;
+
 
     public CourseSchedule(){
         roomList = new ArrayList<>();
@@ -28,30 +34,36 @@ public class CourseSchedule {
         lectureList = new ArrayList<>();
     }
 
-    @ValueRangeProvider(id = "availableRooms")
-    @ProblemFactCollectionProperty
     public List<Integer> getRoomList() {
         return roomList;
     }
 
-    @ValueRangeProvider(id = "availablePeriods")
-    @ProblemFactCollectionProperty
     public List<Integer> getPeriodList() {
         return periodList;
     }
 
-    @PlanningEntityCollectionProperty
     public List<Lecture> getLectureList() {
         return lectureList;
     }
 
-    @PlanningScore
     public HardSoftScore getScore() {
         return score;
     }
 
     public void setScore(HardSoftScore score) {
         this.score = score;
+    }
+
+    public void setRoomList(List<Integer> roomList) {
+        this.roomList = roomList;
+    }
+
+    public void setPeriodList(List<Integer> periodList) {
+        this.periodList = periodList;
+    }
+
+    public void setLectureList(List<Lecture> lectureList) {
+        this.lectureList = lectureList;
     }
 
     public void printCourseSchedule() {
