@@ -13,6 +13,6 @@ public interface CompanyRepository extends JpaRepository<Company, Integer> {
     @Query("SELECT C FROM Company C WHERE C.contactPerson.firstName = ?1")
     List<Company> findByContactPersonFirstNameWithJPQL(String firstName);
 
-    @Query(value = "SELECT * FROM company WHERE contact_first_name = ?1", nativeQuery = true)
+    @Query(value = "SELECT * FROM \"company\" WHERE \"contact_first_name\" = ?1", nativeQuery = true)
     List<Company> findByContactPersonFirstNameWithNativeQuery(String firstName);
 }
