@@ -13,7 +13,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import static org.junit.Assert.*;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest
+@SpringBootTest(classes = AnnotationScannerApp.class)
 public class SampleAnnotationScannerUnitTest {
     @Autowired
     private List<SampleAnnotationScanner> scannerList;
