@@ -12,7 +12,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 @RunWith(SpringRunner.class)
-@SpringBootTest
+@SpringBootTest(classes = OpenApiApplication.class)
 @AutoConfigureMockMvc
 public class OpenApiPetsIntegrationTest {
 
